Add a rAF+transform based dismiss animation.
Because we need to query the current transform when we dismiss
an animation while an existing animation is in progress, we
can't use CSS animations. So, rAF.

I've left in the height-based one because it's a million times
simpler and perhaps the real case we should be optimizing. It's
triggered by loading demo.html?animateHeightOnDismiss.

Also, this all got busted by adding in the message_groups.
We don't properly animate the "when" divs yet.