Do not remove DOM element when dismissing an item.
When dismissing we now delete the item from the fakeData. For doing so
an id is added to all items.

After that we refresh all the items.

Also, make sure items visually after the dismissed item are moved up

The visual order is not necessarily the same as the DOM order as
the children can be split into two groups with different y
transform.